Making the library work with the Arduino IDE. (#13)
* Move includes to the header file.

The header file uses some types from the files included in the source file. Arduino IDE doesn't compile because the header is included before the other files. Moving the includes to the header file removes the need for special include order.

* Include two libraries and move includes from source.

This commit adds two libraries that are required to compile the source file, and also moves libraries from the source file to the header file because their data types are used here.